+++ classes/local_contact.php
@@ -18,13 +18,13 @@
  * This plugin for Moodle is used to send emails through a web form.
  *
  * @package    local_contact
- * @copyright  2016-2022 TNG Consulting Inc. - www.tngconsulting.ca
+ * @copyright  2016-2023 TNG Consulting Inc. - www.tngconsulting.ca
  * @author     Michael Milette
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 /**
  * local_contact class. Handles processing of information submitted from a web form.
- * @copyright  2016-2022 TNG Consulting Inc. - www.tngconsulting.ca
+ * @copyright  2016-2023 TNG Consulting Inc. - www.tngconsulting.ca
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class local_contact {